feat: Rename template to Openresume
